requirements.txt
tensorflow
tqdm
colorama
regex
python-Levenshtein
requests
sh
google-cloud-bigquery
